Add functionality to save file before closing it commpletely

+++ src/renderer/src/composables/ipcListeners/useCloseActiveTab.ts
@@ -5,6 +5,14 @@ export const useCloseActiveTab = () => {
     const FileStore = useFileStore();
     const deactivate = window.electron.ipcRenderer.on('closeActiveTab', () => {
         if (FileStore.areFilesEmpty) return;
+
+        // Check if file is not saved
+        const File = FileStore.getCurrentFile;
+        const HasChanged = File.content !== File.previousContent;
+        if (HasChanged) {
+            window.api.showConfirmDialog();
+            return;
+        }
         FileStore.deleteCurrentActiveFile();
     });
